Fixed loading of x11 and some warnings.

+++ src/br_string_pool.h
@@ -128,7 +128,7 @@ brsp_id_t brsp_new1(brsp_t* sp, int size) {
 }
 
 brsp_id_t brsp_push(brsp_t* sp, br_strv_t sv) {
-  brsp_id_t new_id = brsp_new1(sp, sv.len + 8);
+  brsp_id_t new_id = brsp_new1(sp, (int)sv.len + 8);
   brsp_insert_strv_at_end(sp, new_id, sv);
   return new_id;
 }
